move json parsing logic into JSONUtils
