+++ common-portal/src/main/java/org/ow2/proactive_grid_cloud_portal/common/client/Controller.java
@@ -78,83 +78,7 @@ public abstract class Controller {
      */
     public abstract String getLogo350Url();
 
-    /**
-     * @param throwable a serialized JSON Exception
-     * @return the value of the 'errorMessage' key
-     */
-    public static String getJsonErrorMessage(Throwable throwable) {
-        String msg = throwable.getMessage();
-        return getJsonErrorMessage(msg);
-    }
-
-    /**
-     * @param throwable a serialized JSON Exception
-     * @return the value of the 'httpErrorCode' key, or -1
-     */
-    public static int getJsonErrorCode(Throwable throwable) {
-        String msg = throwable.getMessage();
-        return getJsonErrorCode(msg);
-    }
-
-    /**
-     * @param str String representation of a serialized JSON Exception
-     * @return the value of the 'errorMessage' key
-     */
-    public static String getJsonErrorMessage(String str) {
-        try {
-            JSONObject exc = JSONParser.parseStrict(str).isObject();
-            if (exc != null && exc.containsKey("errorMessage")) {
-                return retrieveErrorMessage(exc);
-            } else if (exc != null && exc.containsKey("exception")) {
-                JSONObject nestedExc = exc.get("exception").isObject();
-                return retrieveErrorMessage(nestedExc);
-            }
-        } catch (Exception e) {
-            if (str != null) {
-                return str;
-            } else {
-                return "<no reason>";
-            }
-        }
-        return null;
-    }
-
-    private static String retrieveErrorMessage(JSONObject exc) {
-        JSONValue val = exc.get("errorMessage");
-        if (val == null || val.isString() == null) {
-            return "<no reason>";
-        } else {
-            return val.isString().stringValue();
-        }
-    }
-
-    /**
-     * @param str String representation of a serialized JSON Exception
-     * @return the value of the 'httpErrorCode' key, or -1
-     */
-    public static int getJsonErrorCode(String str) {
-        try {
-            JSONObject exc = JSONParser.parseStrict(str).isObject();
-            if (exc != null && exc.containsKey("httpErrorCode")) {
-                return retrieveErrorCode(exc);
-            } else if (exc != null && exc.containsKey("exception")) {
-                JSONObject nestedExc = exc.get("exception").isObject();
-                return retrieveErrorCode(nestedExc);
-            }
-        } catch (Exception e) {
-            return -1;
-        }
-        return -1;
-    }
-
-    private static int retrieveErrorCode(JSONObject exc) {
-        JSONValue val = exc.get("httpErrorCode");
-        if (val == null || val.isNumber() == null) {
-            return -1;
-        } else {
-            return (int) val.isNumber().doubleValue();
-        }
-    }
+    
 
     /**
      * Parse a JSON string
