+++ rm-portal/src/main/java/org/ow2/proactive_grid_cloud_portal/rm/client/monitoring/charts/MBeanDetailedView.java
@@ -38,6 +38,7 @@ package org.ow2.proactive_grid_cloud_portal.rm.client.monitoring.charts;
 
 import java.util.List;
 
+import org.ow2.proactive_grid_cloud_portal.common.client.json.JSONUtils;
 import org.ow2.proactive_grid_cloud_portal.common.client.model.LogModel;
 import org.ow2.proactive_grid_cloud_portal.rm.client.RMController;
 import org.ow2.proactive_grid_cloud_portal.rm.client.RMModel;
@@ -126,7 +127,7 @@ public class MBeanDetailedView extends DetailViewer {
                     extraCallback.onFailure(caught);
                 }
 
-                if (RMController.getJsonErrorCode(caught) == 401) {
+                if (JSONUtils.getJsonErrorCode(caught) == 401) {
                     LogModel.getInstance().logMessage("You have been disconnected from the server.");
                 }
             }
