+++ rm-portal/src/main/java/org/ow2/proactive_grid_cloud_portal/rm/client/monitoring/views/FileSystemView.java
@@ -39,6 +39,7 @@ package org.ow2.proactive_grid_cloud_portal.rm.client.monitoring.views;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.ow2.proactive_grid_cloud_portal.common.client.json.JSONUtils;
 import org.ow2.proactive_grid_cloud_portal.common.client.model.LogModel;
 import org.ow2.proactive_grid_cloud_portal.rm.client.RMController;
 import org.ow2.proactive_grid_cloud_portal.rm.client.RMModel;
@@ -157,7 +158,7 @@ public class FileSystemView extends VLayout {
                     }
 
                     public void onFailure(Throwable caught) {
-                        if (RMController.getJsonErrorCode(caught) == 401) {
+                        if (JSONUtils.getJsonErrorCode(caught) == 401) {
                             LogModel.getInstance().logMessage("You have been disconnected from the server.");
                         }
                     }
