+++ rm-portal/src/main/java/org/ow2/proactive_grid_cloud_portal/rm/client/monitoring/views/NetworkView.java
@@ -36,6 +36,11 @@
  */
 package org.ow2.proactive_grid_cloud_portal.rm.client.monitoring.views;
 
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
+
+import org.ow2.proactive_grid_cloud_portal.common.client.json.JSONUtils;
 import org.ow2.proactive_grid_cloud_portal.common.client.model.LogModel;
 import org.ow2.proactive_grid_cloud_portal.rm.client.RMController;
 import org.ow2.proactive_grid_cloud_portal.rm.client.RMModel;
@@ -43,10 +48,6 @@ import org.ow2.proactive_grid_cloud_portal.rm.client.RMServiceAsync;
 import org.ow2.proactive_grid_cloud_portal.rm.client.monitoring.Reloadable;
 import org.ow2.proactive_grid_cloud_portal.rm.client.monitoring.charts.MBeanDetailedView;
 import org.ow2.proactive_grid_cloud_portal.rm.client.monitoring.charts.NetworkDetailedAreaChart;
-
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
 
 import com.google.gwt.json.client.JSONArray;
 import com.google.gwt.json.client.JSONObject;
@@ -161,7 +162,7 @@ public class NetworkView extends VLayout implements Reloadable {
                     }
 
                     public void onFailure(Throwable caught) {
-                        if (RMController.getJsonErrorCode(caught) == 401) {
+                        if (JSONUtils.getJsonErrorCode(caught) == 401) {
                             LogModel.getInstance().logMessage("You have been disconnected from the server.");
                         }
                     }
